Support for resource priority attribute postpone.
This commit adds support for the W3C's draft specification for the
postpone resource priority attribute. Any picture element with postpone
or postpone="true" will not be processed until it is visible on the
screen. Additionally, this commit eliminates the unnecessary loading of
IE8 backup images in browsers that support the full polyfill.